+++ Models/Products/ProductViewModel.cs
@@ -3,7 +3,7 @@ public class ProductViewModel
 {
     public string? name {get;set;}
 
-    public decimal? price {get;set;}
+    public double? price {get;set;}
 
     public string? description {get;set;}
